Add app.ts and basics.ts files, remove counter.ts, and update main.ts
