+++ src/app.ts
@@ -1,0 +1,8 @@
+console.log("app.js loaded");
+
+const person: { name: string; age: number } = {
+  name: "Jerome",
+  age: 36,
+};
+
+console.log(person.name);
